pyproject.toml
[tool.poetry]
name = "astro-bot"
version = "0.1.0"
description = ""
authors = ["igor"]
readme = "README.md"
packages = [{include = "astro_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.25.1"
bs4 = "^0.0.1"
requests = "^2.28.2"
pytz = "^2022.7.1"
python-dotenv = "^1.0.0"
flake8 = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
